+++ Apps/Examples/Examples/All Examples/BuildingExtrusionsExample.swift
@@ -26,7 +26,7 @@ public class BuildingExtrusionsExample: UIViewController, ExampleProtocol {
                                           zoom: 15.5,
                                           bearing: -17.6,
                                           pitch: 45)
-        mapView.camera.setCamera(to: cameraOptions)
+        mapView.mapboxMap.setCamera(to: cameraOptions)
 
         // The below lines are used for internal testing purposes only.
         DispatchQueue.main.asyncAfter(deadline: .now()+5.0) {