+++ Apps/Examples/Examples/All Examples/FitCameraToGeometryExample.swift
@@ -55,7 +55,7 @@ public class FitCameraToGeometryExample: UIViewController, ExampleProtocol {
         }
 
         let newCamera = mapView.mapboxMap.camera(for: polygon, padding: .zero, bearing: 0, pitch: 0)
-        mapView.camera.setCamera(to: newCamera)
+        mapView.mapboxMap.setCamera(to: newCamera)
         // The below line is used for internal testing purposes only.
         self.finish()
     }