+++ Apps/Examples/Examples/All Examples/LineGradientExample.swift
@@ -22,7 +22,7 @@ public class LineGradientExample: UIViewController, ExampleProtocol {
             // Set the center coordinate and zoom level.
             let centerCoordinate = CLLocationCoordinate2D(latitude: 38.875, longitude: -77.035)
             let camera = CameraOptions(center: centerCoordinate, zoom: 12.0)
-            self.mapView.camera.setCamera(to: camera)
+            self.mapView.mapboxMap.setCamera(to: camera)
         }
     }