+++ Sources/MapboxMaps/Gestures/CameraAnimationsManagerProtocol.swift
@@ -6,8 +6,6 @@ internal protocol CameraAnimationsManagerProtocol: AnyObject {
 
     var options: CameraBoundsOptions { get }
 
-    func setCamera(to camera: CameraOptions)
-
     func ease(to camera: CameraOptions,
               duration: TimeInterval,
               curve: UIView.AnimationCurve,